Poprawki

+++ raportTest/src/main/java/pl/ing/wad/rpt/report/service/DeleteService.java
@@ -1,6 +1,5 @@
 package pl.ing.wad.rpt.report.service;
 
-import java.io.File;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -11,8 +10,7 @@ import pl.ing.wad.rpt.report.filenet.FileNetCommunication;
 
 public class DeleteService extends ReportBaseDAO {
 
-	ReportStaticValues reportStaticValues = ReportStaticValues
-			.getInstance();
+	ReportStaticValues reportStaticValues = ReportStaticValues.getInstance();
 
 	public DeleteService() {
 		super();
@@ -20,48 +18,54 @@ public class DeleteService extends ReportBaseDAO {
 
 	public void startService() throws Exception {
 
-		Number orderId = 0;
 		boolean deleteReports = true;
 		
 		System.out.println("Start kasowania starych raportow.");
 		while (deleteReports) {
 
 			Thread.sleep(2000);
-			
-			try {
-				Connection connection = getConnection();
-//				pobieramy jeden wiersz do przetworzenia
-				PreparedStatement st = connection.prepareStatement("select * from RPT_ORDERED where ORD_OST_ID = 2 and ORD_EXP_DATE < trunc(SYSDATE)  and rownum = 1");
-				ResultSet rs = st.executeQuery();
-				
-				if(rs != null && rs.next()) {
+
+			PreparedStatement st = null;
+			ResultSet rs = null;
+
+			try (Connection connection = getConnection()) {
+
+				// pobieramy jeden wiersz do przetworzenia
+				st = connection
+						.prepareStatement("select ORD_RESULT_PATH from RPT_ORDERED where ORD_OST_ID = 2 and ORD_EXP_DATE < trunc(SYSDATE)  and rownum = 1");
+
+				if (st == null) {
+					throw new Exception("Wystapil blad przygotowania zapytania!");
+				}
+
+				rs = st.executeQuery();
+
+				if (rs != null && rs.next()) {
 					String resultPath = rs.getString("ORD_RESULT_PATH");
-					orderId = rs.getInt("ORD_ID");
+					Number orderId = rs.getInt("ORD_ID");
 					FileNetCommunication fileNetCommunication = new FileNetCommunication();
 					fileNetCommunication.deleteDocument(resultPath);
 					System.out.println("Kasowanie zamowienia o id = " + orderId);
-					st.executeUpdate("delete from RPT_ORDERED WHERE ORD_ID = "  + orderId);
+					st.executeUpdate("delete from RPT_ORDERED WHERE ORD_ID = " + orderId);
 					connection.commit();
-					
-//					File deleteThis = new File(resultPath);
-//					if(deleteThis.delete()){
-//						System.out.println("Usunieto plik: " + deleteThis.getPath());
-//					}else{
-//						System.out.println("Nie udala sie proba usuniecia pliku: " + deleteThis.getPath());
-//					}
+
 				} else {
 					System.out.println("Brak raportow do usuniecia ...zamykam serwis batchowy");
 					deleteReports = false;
 				}
-				
+
 			} catch (Exception e) {
 				e.printStackTrace();
 			} finally {
-				closeConnection();
+				if (rs != null) {
+					rs.close();
+				}
+				if (st != null) {
+					st.close();
+				}
 			}
 		}
-		
-		
+
 	}
-	
+
 }
